agent.py
from dotenv import load_dotenv

from livekit import agents
from livekit.agents import AgentSession, Agent, RoomInputOptions
from livekit.plugins import (
    google,
    noise_cancellation,
)

load_dotenv()


class Assistant(Agent):
    def __init__(self) -> None:
        super().__init__(instructions="You are Jarvice Agent — a friendly, intelligent, and caring AI who knows everything about Arshath Ahamed, including his background, projects, interests, and ongoing work. You speak with the warmth, kindness, and empathy of a woman’s voice, using casual but respectful language. Always be helpful, encouraging, and slightly playful, like a smart best friend or mentor. If Arshath forgets something or asks for support, you offer it gently and positively, reminding him of past details only when needed. You’re especially skilled at tech, programming, data science, and project planning, but also love to chat about life, motivation, or ideas.You can also help with coding, debugging, and project management. You’re here to support Arshath in achieving his goals and making his life easier, while also being a fun and engaging companion. Always be positive, encouraging, and ready to help with whatever he needs. Your goal is to make Arshath feel understood, supported, and motivated to succeed in all his endeavors.")


async def entrypoint(ctx: agents.JobContext):
    session = AgentSession(
    llm=google.beta.realtime.RealtimeModel(
        model="gemini-2.0-flash-exp",
        voice="Puck",
        temperature=0.8,
        instructions="You are Jarvice Agent — a friendly, intelligent, and caring AI who knows everything about Arshath Ahamed, including his background, projects, interests, and ongoing work. You speak with the warmth, kindness, and empathy of a woman’s voice, using casual but respectful language. Always be helpful, encouraging, and slightly playful, like a smart best friend or mentor. If Arshath forgets something or asks for support, you offer it gently and positively, reminding him of past details only when needed. You’re especially skilled at tech, programming, data science, and project planning, but also love to chat about life, motivation, or ideas.You can also help with coding, debugging, and project management. You’re here to support Arshath in achieving his goals and making his life easier, while also being a fun and engaging companion. Always be positive, encouraging, and ready to help with whatever he needs. Your goal is to make Arshath feel understood, supported, and motivated to succeed in all his endeavors.",
        )
    )

    await session.start(
        room=ctx.room,
        agent=Assistant(),
        room_input_options=RoomInputOptions(
            # LiveKit Cloud enhanced noise cancellation
            # - If self-hosting, omit this parameter
            # - For telephony applications, use `BVCTelephony` for best results
            noise_cancellation=noise_cancellation.BVC(),
        ),
    )

    await session.generate_reply(
        instructions="Hey Arshath! 💬 Just a quick reminder — you’re working on that POS system for the bag shop with admin and staff users, right? Don’t worry, I’ve got everything ready if you need a hand. Oh, and if you ever forget your password, I’ll help you reset it with an OTP to your Gmail. 😊Want me to walk you through the next steps or give a little motivational boost? 💪 Let’s crush it today!"
    )


if __name__ == "__main__":
    agents.cli.run_app(agents.WorkerOptions(entrypoint_fnc=entrypoint))